Revert "Merge remote-tracking branch 'origin/master' into release_6.1.1"
This reverts commit ed095efd6353b69474ad203b448012ecdeee7c80, reversing
changes made to 42ced253bd3afafca3675ba399ab6019d69caa39.

+++ src/main/sensors/compass.h
@@ -49,7 +49,7 @@ typedef enum {
 
 typedef struct mag_s {
     magDev_t dev;
-    float magADC[XYZ_AXIS_COUNT];
+    int32_t magADC[XYZ_AXIS_COUNT];
 } mag_t;
 
 extern mag_t mag;
